backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "fastapi>=0.122.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-text-splitters>=1.0.0",
    "langchain-xai>=1.1.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "pypdf>=6.4.0",
    "python-multipart>=0.0.20",
    "supabase>=2.24.0",
]
